requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0
numpy>=1.24.0
PyOpenGL>=3.1.6
PyOpenGL_accelerate>=3.1.6
glfw>=2.6.0
pyrr>=0.10.3
